requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
numpy==1.26.0
pillow==10.0.0
tqdm==4.66.1
requests==2.31.0
pytube==12.1.0
ffmpeg-python==0.4.0
pydub==0.25.1
faster-whisper==0.6.1
openai==1.11.0
transformers==4.35.0
torch==2.2.0
timm==0.9.0
scikit-learn==1.3.0
scipy==1.11.3
opencv-python-headless==4.8.0
